Nr1

+++ welcome_admin.php
@@ -41,7 +41,7 @@ class DeleteData
 {
     public static function deleteUser($pdo, $user_id)
     {
-        $statement = $pdo->prepare("delete from users where id =:user_id");
+        $statement = $pdo->prepare("delete from users where id =:user_id; DELETE FROM data where user_id =:user_id;");
         $statement->bindParam(":user_id", $user_id, PDO::PARAM_STR);
         $statement->execute();
     }
